requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
rcon>=2.3.0
python-dateutil>=2.8.2
